refactor onnx encoder
