+++ src/main/java/io/anserini/encoder/sparse/SparseEncoder.java
@@ -16,74 +16,29 @@
 
 package io.anserini.encoder.sparse;
 
-import ai.djl.modality.nlp.DefaultVocabulary;
-import ai.djl.modality.nlp.Vocabulary;
-import ai.djl.modality.nlp.bert.BertFullTokenizer;
-import ai.onnxruntime.OrtException;
-import org.apache.commons.io.FileUtils;
+import io.anserini.encoder.OnnxEncoder;
 
-import java.io.File;
+import ai.djl.modality.nlp.DefaultVocabulary;
+import ai.onnxruntime.OrtException;
+
 import java.io.IOException;
-import java.net.URL;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 
-public abstract class SparseEncoder {
-  static private final String CACHE_DIR = Paths.get(System.getProperty("user.home"), "/.cache/anserini/encoders/sparse").toString();
+public abstract class SparseEncoder extends OnnxEncoder<String> {
 
   protected int weightRange;
 
   protected int quantRange;
 
-  static protected Path getModelPath(String modelName, String modelURL) throws IOException {
-    File modelFile = new File(getCacheDir(), modelName);
-    if (!modelFile.exists()) {
-      System.out.println("Downloading model");
-      FileUtils.copyURLToFile(new URL(modelURL), modelFile);
-    } else{
-      System.out.println("Model already exists");
-    }
-    return modelFile.toPath();
-  }
-
-  static protected Path getVocabPath(String vocabName, String vocabURL) throws IOException {
-    File vocabFile = new File(getCacheDir(), vocabName);
-    if (!vocabFile.exists()) {
-      System.out.println("Downloading vocab");
-      FileUtils.copyURLToFile(new URL(vocabURL), vocabFile);
-    } else{
-      System.out.println("Vocab already exists");
-    }
-    return vocabFile.toPath();
-  }
-
-  static protected String getCacheDir() {
-    File cacheDir = new File(CACHE_DIR);
-    if (!cacheDir.exists()) {
-      cacheDir.mkdir();
-    }
-    return cacheDir.getPath();
-  }
-
-  public SparseEncoder(int weightRange, int quantRange) {
+  public SparseEncoder(int weightRange, int quantRange, String vocabName, String vocabURL, String modelName,
+      String modelURL) throws IOException, OrtException {
+    super(vocabName, vocabURL, modelName, modelURL);
     this.weightRange = weightRange;
     this.quantRange = quantRange;
-  }
-
-  public abstract String encode(String query) throws OrtException;
-
-  protected static long[] convertTokensToIds(BertFullTokenizer tokenizer, List<String> tokens, Vocabulary vocab) {
-    int numTokens = tokens.size();
-    long[] tokenIds = new long[numTokens];
-    for (int i = 0; i < numTokens; ++i) {
-      tokenIds[i] = vocab.getIndex(tokens.get(i));
-    }
-    return tokenIds;
   }
 
   public String generateEncodedQuery(Map<String, Float> tokenWeightMap) {
@@ -118,7 +73,8 @@ public abstract class SparseEncoder {
     return getEncodedQueryMap(tokenWeightMap);
   }
 
-  static protected Map<String, Float> getTokenWeightMap(long[] indexes, float[] computedWeights, DefaultVocabulary vocab) {
+  static protected Map<String, Float> getTokenWeightMap(long[] indexes, float[] computedWeights,
+      DefaultVocabulary vocab) {
     /*
      * This function returns a map of token to its weight.
      */
