+++ src/test/java/io/anserini/encoder/SparseEncoderInferenceTest.java
@@ -1,0 +1,10 @@
+package io.anserini.encoder;
+
+public abstract class SparseEncoderInferenceTest extends EncoderInferenceTest{
+
+    public SparseEncoderInferenceTest(String modelName, String modelUrl, Object[][] examples) {
+        super(modelName, modelUrl, examples);
+    }
+    
+    
+}
